api/requirements.txt
-f https://download.pytorch.org/whl/torch_stable.html
certifi==2022.6.15
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.5
Flask==2.1.2
gunicorn==20.1.0
idna==3.3
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
numpy==1.23.0
Pillow==9.1.1
requests==2.28.0
timm==0.5.4
torch==1.11.0+cpu
torchvision==0.12.0+cpu
typing-extensions==4.2.0
urllib3==1.26.9
Werkzeug==2.1.2
zipp==3.8.0